requirements.txt
backports.zoneinfo>=0.2.1; python_version < "3.9"
tzdata>=2024.1
datasette>=0.65
streamlit>=1.38
plotly>=5.20
statsmodels>=0.14
